Add checkstyle
