+++ src/Test/java/duke/DeadlineTest.java
@@ -1,7 +1,10 @@
-package Duke;
-import org.junit.jupiter.api.Test;
-import java.time.LocalDateTime;
+package duke;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.time.LocalDateTime;
+
+import org.junit.jupiter.api.Test;
 
 /**
  * JUnit test for Deadline class.
