+++ src/Test/java/duke/StorageTest.java
@@ -1,7 +1,10 @@
-package Duke;
-import org.junit.jupiter.api.Test;
-import java.io.IOException;
+package duke;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.io.IOException;
+
+import org.junit.jupiter.api.Test;
 
 /**
  * JUnit test for Storage class.
