+++ src/main/java/duke/Event.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 /**
  * Represents an event task.
@@ -26,8 +26,8 @@ public class Event extends Task {
      */
     @Override
     public String toString() {
-        return "[E]" +
-                super.toString() + " (from: " + from + " to: " + to + ")";
+        return "[E]"
+                + super.toString() + " (from: " + from + " to: " + to + ")";
     }
 
     /**
@@ -36,7 +36,7 @@ public class Event extends Task {
      */
     @Override
     public String toSaveString() {
-        return "E | " + (isDone ? "1" : "0") + " | " +
-                description + " | " + from + " - " + to;
+        return "E | " + (isDone ? "1" : "0") + " | "
+                + description + " | " + from + " - " + to;
     }
 }
