+++ src/main/java/duke/Todo.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 /**
  * Represents a todo task.
@@ -28,7 +28,6 @@ public class Todo extends Task {
      */
     @Override
     public String toSaveString() {
-        return "T | " +
-                (isDone ? "1" : "0") + " | " + description;
+        return "T | " + (isDone ? "1" : "0") + " | " + description;
     }
 }
